Implement Merkle Tree-level Hashing methods
Implements the Poseidon Hash for a Merkle Tree
level for both `Scalar` and `Gadget` strategies.

This relies on `perm` function exported from `Hades252`.
It includes the logic required by the Poseidon Hash
which computes the bitflags in respect of the presence of
the level leafs.

Actually, this functions support the merkle-level hashing
for `Scalar` & `Variable`.

The maximum ammount of leafs per level "ARITY" is actually
4. So for bigger trees, this functions will not behave
correctly. This is a tradeoff we need to do until we can
support custom WIDTH permutations in Hades252 repo.