Implement proof module for Poseidon252
This comes with a `merkle_opening_gadget` fn which
allows to: Provided a `&mut StandardComposer` &
a `PoseidonBranch`, generate a merkle oppening proof
inside the `StandardComposer` provided.

The root of the tree is used as a Public Input so there's
no way for the `Prover` to cheat with the data he/she
provides.