Refactored hash module w/ bitflag dependant fn's
The goal of this library is to provide a way to easily
generate Merkle Tree Proofs inside of ZK Constraint Systems.

But to grant some level of freedom to the user, we've splited
the implementation of the merkle-level hashing logic in 4 functions:

- Implemented Scalar Merkle-tree-level hashing for levels with and
without the bitflags coefficient.

- Implemented Variable Merkle-Tree-level hashing gadget for levels with
and without the bitflags coefficient.

Closes #9