Implement Merkle-Tree Opening proof gen gadget
Provided a `PoseidonBranch`, this method takes a `&mut StandardComposer`
and performs the addition of the needed constraints into the Composer
so it can demonstrate that the path to the root is correct.

- Root is used as a PublicInput so the `Verifier` can check that
indeed he/she is verifying the path to the correct root and de-facto
the correct tree.

This partially closes #10 since we need to add also a `Scalar` proving
method.