Fix kelvin tree offset logic for branch levels
We were considering offset the position of the hash of a level
on the upper level when indeed it was the position of the hash
of the previous level on the actual level.

If we're on the bottom level, it references the leaf we're proving
inclusion for.

- Refactored `PoseidonBranch::from(&Branch)` method to work according
to the right logic that offsets have in kelvin.
- Rename `upper_lvl_hash` as `offset`.
- Refactor Scalar and Variable Proofs with the new offset logic.
- Implement tests for `Scalar` backend and `ZK::Variable` backend
that check the consistency of the branch propperly.

Closes #10 #13 & #9