Port Poseidon repo to be Canon-compatible
We need to be able to implement/derive `Canon` for the trees
that we use in our Smart Contracts. Therefore this crate needed a
refactor so that the Annotations and Tree structures that it exports
carry the Canon traits implemented propperly.

- Removed legacy trait implementations such as kelvin::Content
, ByteHash & others which have been deprecated.
- Enforce that `PoseidonTree` and `PoseidonBranch` store types
which implement `Canon<canonical::Store>`
- Refactor all of the trait implementations to have `Canon` implemented.
- Fixed & refactored the `extend_storage_scalar` macro so that it derives
`Canon`.
- Refactored the error handling to use correctly the
`Store::InvalidEncoding` cannonical error type.

Closes #79