Sponge hash defined as a no-std function.
The only thing preventing the sponge hash of being no-std was the
padding, that used a variable sized vector.

This is not required. If we have a fitting strategy to pad without
recreating the message as a new vector, and being aware of the required
modifications on the last chunk, then we can achieve the same result
without creating a vector.