Merkle opening constant circuit description
PLONK requires the circuit description + witness indexes to be constant
so the same verifier data can be reused to check different proofs.

The previous implementation contained an optimization that swaps the
witness index to reduce the number of gates.

This optimization may lead to inconsistencies when reusing verifier keys
because the verification will depend on the disposition of the witness
data, and this is undesirable.

Resolves #122